--- src/utils/sleep.ts
@@ -1,3 +1,0 @@
-export function sleep(milliseconds: number) {
-    return new Promise((resolve) => setTimeout(resolve, milliseconds));
-}
